feat: use internal IPFS gateway for project detail page
